// protocol-designer/src/collision-types.d.ts
export interface DragRect {
    xStart: number;
    yStart: number;
    xDynamic: number;
    yDynamic: number;
}
export interface GenericRect {
    x0: number;
    x1: number;
    y0: number;
    y1: number;
}
export interface BoundingRect {
    x: number;
    y: number;
    width: number;
    height: number;
}